requirements.txt
# API Clients and Core Dependencies
requests>=2.31.0
PyGithub>=2.1.1
huggingface-hub>=0.19.0

# Database
pymongo>=4.6.0
dnspython>=2.4.2

# Testing
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Utilities
python-dotenv>=1.0.0
dataclasses>=0.6
typing-extensions>=4.8.0

# Data Processing
beautifulsoup4>=4.12.2
lxml>=4.9.3
PyYAML>=6.0.1

# Optional: For advanced summarization
transformers>=4.35.2
torch>=2.1.0
sentencepiece>=0.1.99

# Development Tools
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0